Isolate pool and labeled managers
